Update Car.
Created Car DTO to show that if could be created directly from a query in Repo.

Showed that setters could be used in Constructors. but thats debated if it's a good practice or not..

Also Created a Controller for Car to demo what a response could look like. From Postman.